test(category-controller): create tests for delete method

+++ w4/s3/controllers/category_controller.rb
@@ -10,6 +10,11 @@ class CategoryController
 
   def self.delete(params)
     category = Category.get_by_id(params["id"])
+
+    if category.nil? then
+      return
+    end
+
     category.delete()
   end
